+++ src/components/services/EdgeCoreManager.tsx
@@ -9,7 +9,7 @@ import makeMoneroIo from 'edge-currency-monero/lib/react-native-io'
 import { debugUri as currencyPluginsDebugUri, makePluginIo as makeCurrencyPluginsIo, pluginUri as currencyPluginsUri } from 'edge-currency-plugins'
 import { debugUri as exchangeDebugUri, pluginUri as exchangeUri } from 'edge-exchange-plugins'
 import * as React from 'react'
-import RNBootSplash from 'react-native-bootsplash'
+import BootSplash from 'react-native-bootsplash'
 import { getBrand, getDeviceId } from 'react-native-device-info'
 
 import { ENV } from '../../env'
@@ -129,7 +129,7 @@ export function EdgeCoreManager(props: Props) {
   function hideSplash() {
     if (!splashHidden.current) {
       setTimeout(() => {
-        RNBootSplash.hide({ fade: true }).catch(err => showError(err))
+        BootSplash.hide({ fade: true }).catch(err => showError(err))
       }, 200)
       splashHidden.current = true
     }